requirements.txt
aiogram
python-dotenv
requests
google_drive_downloader
